+++ Angular/src/app/app-routing.module.ts
@@ -33,6 +33,7 @@ import { EmployeeDetailsComponent } from './employee/employee/viewEmployee/view-
 //Broker
 import { BrokerComponent } from './broker/broker/broker.component';
 import { ViewBrokerComponent } from './broker/broker/viewBroker/view-broker/view-broker.component';
+import { GenerateBrokerGraphComponent } from './broker/broker/generateBrokerGraph/generate-broker-graph/generate-broker-graph.component';
 import { GenerateBrokerReportComponent } from './broker/broker/generateBrokerReport/generate-broker-report/generate-broker-report.component';
 //Calendar
 import { CalendarComponent } from './calendar/calendar/calendar.component';
@@ -129,6 +130,14 @@ const routes: Routes = [
   {
     path: 'screensaver',
     component: ScreensaverComponent,
+    canActivate: [RoleGuard],
+    data: {
+      expectedRole: 'Admin',
+    },
+  },
+  {
+    path: 'brokerGraph',
+    component: GenerateBrokerGraphComponent,
     canActivate: [RoleGuard],
     data: {
       expectedRole: 'Admin',
