+++ Angular/src/app/app.module.ts
@@ -143,6 +143,7 @@ import { CalculatorComponent } from './calculator/calculator/calculator.componen
 //TIMEOUT
 import { UserActivity } from './services/userActivity.service';
 import { ViewCalendarComponent } from './calendar/calendar/viewCalendar/view-calendar/view-calendar.component';
+import { GenerateBrokerGraphComponent } from './broker/broker/generateBrokerGraph/generate-broker-graph/generate-broker-graph.component';
 
 export function tokenGetter() {
   return localStorage.getItem('Token');
@@ -257,6 +258,7 @@ export function jwtOptionsFactory() {
     SignaturePadComponent,
     CalculatorComponent,
     ViewCalendarComponent,
+    GenerateBrokerGraphComponent,
   ],
   imports: [
     BrowserModule,
