+++ Angular/src/app/broker/broker/generateBrokerGraph/generate-broker-graph/generate-broker-graph.component.html
@@ -1,0 +1,3 @@
+<h1>Broker Graph</h1>
+<br>
+<canvas id="commissionPieChart" width="400" height="400"></canvas>
