+++ Angular/src/app/broker/broker/generateBrokerReport/generate-broker-report/generate-broker-report.component.html
@@ -6,9 +6,15 @@
         Download PDF
     </button>
 </div>
+
 <br>
 <br>
 
 <h1>Broker Graph</h1>
-<br>
-<canvas id="commissionPieChart" width="400" height="400"></canvas>
+
+<div class="download">
+    <button mat-fab extended color="primary" (click)="graph()">
+        <mat-icon>equalizer</mat-icon>
+        View Graph
+    </button>
+</div>
