+++ Angular/src/index.html
@@ -22,6 +22,9 @@
   <!--Map Stuff-->
   <script type="text/javascript" src="https://www.bing.com/api/maps/mapcontrol?key=Av4Br0qLtt-d3hpzzGqNtuwdJO0TgQcX9GbSJBp1dOOFpXdN38VoXkDrGasvYrpo" async defer></script>
 
+  <!--Graph stuff-->
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/3.6.0/chart.min.js"></script>
+
 </head>
 <body class="mat-typography">
   <app-root></app-root>
